odysseus/concurso_novo/requirements.txt
crewai>=0.11.0
python-dotenv>=0.19.0
langchain>=0.1.0
pypdf2>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
spacy>=3.7.2
